FEAT: Modularise Personal Configuration Files
Instead of tracking all configurations in a single source file,
the `./lua/thekpaul/` subdirectory will be used to contain modularised
files for specific purposes.
Currently, three files are created, each for options, keymaps and
plugin manager bootstrapping.
The order in which these files are sourced is explicitly specified in
the overarching `./init.lua` file, which will only be used as an entry
point for configurations without additionally specified operations.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>