FEAT: Lazy.nvim Plugin Manager Bootstrapping
Installs (if necessary) and bootstraps the Lazy.nvim plugin manager
after searching for and checking the plugin configuration modules
directory to determine if plugin manager bootstrapping is necessary.
If issues are found during the directory search and check stage,
this module prouduces a warning message and gracefully returns without
terminating or otherwise producing errors.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>