CHORE: Disable All Neovim Plugins in VS Code
Uses the `vim.g.vscode` variable provided in VS Code environments by the
VSCode-Neovim plugin to detect if Neovim is attached to VS Code.
If it is, then suppress all Neovim plugins managed by Lazy.nvim.

This is a temporary operation to ensure that Neovim plugins do not mess
with VS Code sessions.
Future commits on this branch will be more granular in deciding which
plugins to allow in VS Code (if any), as well as setting different
options and/or keymaps in VS Code.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>